Add 'UnblockUser' and 'UnblockUserComplete' hooks to Special:Unblock
Currently, there are two hooks ('BlockIP' and 'BlockIPComplete') that run when a user or IP is blocked in Special:Block, but no similar hooks when a user is unblocked in Special:Unblock. Adding these two hooks allows an extension to add and remove criteria about a user/IP unblock before it occurs and for possible action after an unblock is completed.

Bug: T50546

+++ includes/specials/SpecialUnblock.php
@@ -209,10 +209,17 @@ class SpecialUnblock extends SpecialPage {
 			return [ 'unblock-hideuser' ];
 		}
 
+		$reason = [ 'hookaborted' ];
+		if ( !Hooks::run( 'UnblockUser', [ &$block, &$performer, &$reason ] ) ) {
+			return $reason;
+		}
+
 		# Delete block
 		if ( !$block->delete() ) {
 			return [ [ 'ipb_cant_unblock', htmlspecialchars( $block->getTarget() ) ] ];
 		}
+
+		Hooks::run( 'UnblockUserComplete', [ $block, $performer ] );
 
 		# Unset _deleted fields as needed
 		if ( $block->mHideName ) {
